type inconsistencies fix
