+++ frontend/app/(auth)/reset-password/page.tsx
@@ -107,7 +107,7 @@ export default function ResetPasswordPage() {
     <div className="flex min-h-screen flex-col items-center justify-center bg-background px-4 py-12">
       <div className="mx-auto flex w-full flex-col justify-center space-y-6 sm:w-[400px]">
         <div className="flex flex-col space-y-2 text-center">
-          <h1 className="text-3xl font-bold tracking-tight">drAIve</h1>
+          <h1 className="text-3xl font-bold tracking-tight">Archyx AI</h1>
           <p className="text-sm text-muted-foreground">
             {step === "request" ? "Reset your password" : "Create new password"}
           </p>
