+++ frontend/app/(dashboard)/layout.tsx
@@ -4,7 +4,7 @@ import { SidebarProvider } from "@/components/ui/sidebar"
 import { AppSidebar } from "@/components/app-sidebar"
 import { SidebarInset } from "@/components/ui/sidebar"
 import { Header } from "@/components/layout/header"
-import { ChatProvider } from "@/components/providers/chat-provider"
+import { ChatProvider } from "@/hooks/chat-provider"
 
 export default function DashboardLayout({ children }: { children: React.ReactNode }) {
   return (
