+++ frontend/components/chat/message-bubble.tsx
@@ -27,8 +27,8 @@ import {
 } from "@/components/ui/dropdown-menu"
 import { Avatar, AvatarFallback, AvatarImage } from "@/components/ui/avatar"
 import { Badge } from "@/components/ui/badge"
-import { useChat, type Message } from "@/components/providers/chat-provider"
-import { useSettings } from "@/components/providers/settings-provider"
+import { useChat, type Message } from "@/hooks/chat-provider"
+import { useSettings } from "@/hooks/settings-provider"
 import { formatDistanceToNow, format } from "date-fns"
 import { cn } from "@/lib/utils"
 import { useToast } from "@/hooks/use-toast"
