+++ frontend/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "drAIve",
+  "name": "archyx-ai",
   "version": "0.1.0",
   "private": true,
   "scripts": {
